refactor: move HistoryListPanelCol interface to centralized type (#2379)
* refactor(HistoryListEntryMaintenance): simplify class binding by removing computed property


* refactor(HistoryListPanel): update import paths for HistoryListPanelCol to use centralized types


---------
