+++ src/components/panels/History/HistoryListEntryMaintenance.vue
@@ -2,7 +2,7 @@
     <tr
         :key="item.id"
         v-longpress:600="(e) => showContextMenu(e)"
-        :class="cssClasses"
+        class="file-list-cursor user-select-none"
         @contextmenu="showContextMenu($event)"
         @click="detailsDialogBool = true">
         <td class="pr-0">
@@ -71,7 +71,7 @@ import {
     mdiNotebookCheck,
     mdiTextBoxSearch,
 } from '@mdi/js'
-import { HistoryListPanelCol } from '@/components/panels/HistoryListPanel.vue'
+import { HistoryListPanelCol } from '@/store/server/history/types'
 import { GuiMaintenanceStateEntry } from '@/store/gui/maintenance/types'
 import HistoryListPanelDetailMaintenance from '@/components/dialogs/HistoryListPanelDetailMaintenance.vue'
 
@@ -94,12 +94,6 @@ export default class HistoryListPanel extends Mixins(BaseMixin) {
     @Prop({ type: Object, required: true }) readonly item!: GuiMaintenanceStateEntry
     @Prop({ type: Array, required: true }) readonly tableFields!: HistoryListPanelCol[]
     @Prop({ type: Boolean, required: true }) readonly isSelected!: boolean
-
-    get cssClasses() {
-        let output = ['file-list-cursor', 'user-select-none']
-
-        return output
-    }
 
     get restFilament() {
         const start = this.item?.start_filament ?? 0
